+++ Sources/Sentry/SentrySDK.m
@@ -1,4 +1,5 @@
 #import "SentrySDK.h"
+#import "SentryAppStartMeasurement.h"
 #import "SentryBreadcrumb.h"
 #import "SentryClient+Private.h"
 #import "SentryCrash.h"
@@ -19,6 +20,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 static SentryHub *currentHub;
 static BOOL crashedLastRunCalled;
+static SentryAppStartMeasurement *appStartMeasurement;
 
 + (SentryHub *)currentHub
 {
@@ -51,6 +53,16 @@ static BOOL crashedLastRunCalled;
 + (void)setCrashedLastRunCalled:(BOOL)value
 {
     crashedLastRunCalled = value;
+}
+
++ (SentryAppStartMeasurement *)appStartMeasurement
+{
+    return appStartMeasurement;
+}
+
++ (void)setAppStartMeasurement:(SentryAppStartMeasurement *)value
+{
+    appStartMeasurement = value;
 }
 
 + (void)startWithOptions:(NSDictionary<NSString *, id> *)optionsDict