+++ Sources/Sentry/SentryTransaction.m
@@ -1,9 +1,11 @@
 #import "SentryTransaction.h"
+#import "NSDictionary+SentrySanitize.h"
 #import "SentryEnvelopeItemType.h"
 
 @implementation SentryTransaction {
     id<SentrySpan> _trace;
     NSArray<id<SentrySpan>> *_spans;
+    NSMutableDictionary<NSString *, id> *measurements;
 }
 
 - (instancetype)initWithTrace:(id<SentrySpan>)trace children:(NSArray<id<SentrySpan>> *)children
@@ -14,8 +16,14 @@
         _trace = trace;
         _spans = children;
         self.type = SentryEnvelopeItemTypeTransaction;
+        measurements = [NSMutableDictionary new];
     }
     return self;
+}
+
+- (void)setMeasurementValue:(id)value forKey:(NSString *)key
+{
+    measurements[key] = value;
 }
 
 - (NSDictionary<NSString *, id> *)serialize
@@ -35,6 +43,10 @@
     }
     mutableContext[@"trace"] = [_trace serialize];
     [serializedData setValue:mutableContext forKey:@"contexts"];
+
+    if (measurements.count > 0) {
+        serializedData[@"measurements"] = [measurements.copy sentry_sanitize];
+    }
 
     return serializedData;
 }