+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 7.1.0
 
+- feat: Measure app start time (#1111)
 - fix: Remove SentryUnsignedLongLongValue (#1118)
 - feat: Expose SentryDebugImageProvider (#1094)
 - docs: Improve code doc on start and endSession (#1098)